requirements.txt
numpy
pandas
streamlit
matplotlib
pillow
scikit-learn
